// src/client/src/app/Modelos.ts
export class Modelos {
  _id?: string;
  Name: string;
  LastName: string;
  NickName: string;
  Age: number;
  BornDate: Date;
  Telephone: string;
  Residence: string;
  Nationality:string;
  Profession:string;
  Stature: string;
  Languages: string;
  Meausurements: Object;
  CostPerHour: number;
  Hotels: boolean;
  Departaments: boolean;
  Travels	: boolean;
  AgencyName: string;
  AgencyUrl: Object;
  Services: Object;
  isDone: boolean;
  n?: number;
}